Parameterize setting up the context
